Rename experimentCalculations -> legacyExperimentCalculations
