+++ frontend/src/scenes/experiments/ExperimentView/Overview.tsx
@@ -7,7 +7,7 @@ import {
 } from '~/queries/schema/schema-general'
 import { ExperimentIdType } from '~/types'
 
-import { getHighestProbabilityVariant, getIndexForVariant } from '../experimentCalculations'
+import { getHighestProbabilityVariant, getIndexForVariant } from '../legacyExperimentCalculations'
 import { experimentLogic } from '../experimentLogic'
 import { VariantTag } from './components'
