+++ frontend/src/scenes/experiments/legacyExperimentCalculations.tsx
